refactor: unify local and GHA unit test runners into single script
- Create unified script/run-unit-tests.js that works in both environments
- Auto-detect GitHub Actions via GITHUB_ACTIONS env var
- Conditionally load @actions/core only when in GHA
- Use 8GB memory limit for GHA, 4GB for local development
- Add --changed-only and --full-suite options
- Update help documentation with new options
- Remove script/run-unit-tests-local.js
- Remove script/github-actions/run-unit-tests-gha.js
- Update package.json to point both test:unit and test:unit:gha to unified script
