+++ package.json
@@ -63,8 +63,8 @@
     "test:coverage-app": "node script/run-app-coverage.js",
     "test:coverage-apps": "npm run test:coverage && node script/app-coverage-report.js",
     "test:coverage-only": "NODE_ENV=test nyc report --all --reporter=json-summary",
-    "test:unit": "node ./script/run-unit-tests-local.js",
-    "test:unit:gha": "node ./script/github-actions/run-unit-tests-gha.js",
+    "test:unit": "node ./script/run-unit-tests.js",
+    "test:unit:gha": "node ./script/run-unit-tests.js",
     "test:unit:summary": "./script/test-summary.js",
     "test:watch": "node script/watch-tests.js",
     "update:schema": "./script/update-json-schema.sh",
