+++ script/run-unit-test-help.js
@@ -4,6 +4,8 @@ const commandLineUsage = require('command-line-usage');
 const helpSections = [
   {
     header: 'Unit test run script',
+    content:
+      'Unified test runner for vets-website. Works in both local development and GitHub Actions environments.',
   },
   {
     header: 'Examples',
@@ -26,6 +28,14 @@ const helpSections = [
         example:
           '$ yarn test:unit --app-folder personalization --coverage --coverage-html',
       },
+      {
+        desc: '5. Run only tests for changed files (like CI)',
+        example: '$ yarn test:unit --changed-only',
+      },
+      {
+        desc: '6. Run full test suite explicitly',
+        example: '$ yarn test:unit --full-suite',
+      },
     ],
   },
   {
@@ -42,20 +52,19 @@ const helpSections = [
         name: 'coverage',
         typeLabel: '{underline boolean}',
         description:
-          'Runs the unit tests with code coverage metrics, and outputs the results to an json report in coverage/',
+          'Runs the unit tests with code coverage metrics, and outputs the results to a json report in coverage/',
       },
       {
         name: 'coverage-html',
         typeLabel: '{underline boolean}',
         description:
-          'Used in conjunction with `--coverage` Runs the unit tests with code coverage metrics, and outputs the results to an html report in coverage/ instead of json.',
+          'Used in conjunction with `--coverage`. Runs the unit tests with code coverage metrics, and outputs the results to an html report in coverage/ instead of json.',
       },
       {
         name: 'app-folder',
         typeLabel: '{underline folder name}',
         description:
           'Run all tests in the specified folder in src/applications',
-        defaultOption: true,
       },
       {
         name: 'path',
@@ -65,9 +74,30 @@ const helpSections = [
         defaultOption: true,
       },
       {
+        name: 'config',
+        typeLabel: '{underline path}',
+        description: 'Path to mocha config file. Defaults to config/mocha.json',
+      },
+      {
+        name: 'reporter',
+        typeLabel: '{underline name}',
+        description: 'Mocha reporter to use (e.g., spec, dot, nyan)',
+      },
+      {
+        name: 'full-suite',
+        typeLabel: '{underline boolean}',
+        description: 'Run the complete test suite (all tests)',
+      },
+      {
+        name: 'changed-only',
+        typeLabel: '{underline boolean}',
+        description:
+          'Run only tests for changed files (auto-detected from CHANGED_FILES env var). This is the default behavior in GitHub Actions.',
+      },
+      {
         name: 'help',
         typeLabel: '{underline boolean}',
-        description: 'Show the usage guide',
+        description: 'Show this usage guide',
       },
     ],
   },
